requirements.txt
# TraderClaw - Automated Trading System
# Core dependencies

# AI / LLM Models
# Gemini (Data gathering with Google Search) - new SDK
google-genai>=1.0.0

# Old SDK (deprecated) - do not use
# google-generativeai>=0.7.0

# DeepSeek (Default strategy model)
# Uses OpenAI-compatible API, no special package needed

# Anthropic Claude (Alternative strategy model)
anthropic>=0.28.0

# OpenAI (For GPT strategy model or DeepSeek compatibility)
openai>=1.30.0

# Data Sources
yfinance>=0.2.28
pycoingecko>=3.1.0
requests>=2.31.0

# Brokers
alpaca-trade-api>=3.0.0
alpaca-py>=0.13.0

# Configuration
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Async support
aiohttp>=3.8.0
aiofiles>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
responses>=0.23.0

# Utilities
python-dateutil>=2.8.2
websocket-client>=1.6.0

# Logging
structlog>=23.0.0
